Extract contextual binding tests
